Create nmap_port_scan.py
Basic python script to use nmap. Utilizes target IP and scan type arguments. I plan to add in additional functions in the script to filter for open, closed, and specific ports for reconnaissance activities. I would also like it to export the results to a text document to be tied to their respective customers/clients/investigations.